Move helper functions to helper module in hope of fixing diff.
